+++ lib/src/components/tab_view.dart
@@ -23,6 +23,7 @@ class TenorTabView extends StatefulWidget {
   final Widget Function(BuildContext, Widget?)? builder;
   final TenorCategoryStyle categoryStyle;
   final Tenor client;
+  final String featuredCategory;
   final int mediaWidth;
   final bool? keepAliveTabView;
   final Future<TenorResponse?> Function(
@@ -40,13 +41,14 @@ class TenorTabView extends StatefulWidget {
     required this.mediaWidth,
     this.builder,
     this.categoryStyle = const TenorCategoryStyle(),
+    String? featuredCategory,
     this.keepAliveTabView,
     this.onLoad,
     this.onSelected,
     this.showCategories = false,
     this.style = const TenorTabViewStyle(),
     super.key,
-  });
+  }) : featuredCategory = featuredCategory ?? '📈 Featured';
 
   @override
   State<TenorTabView> createState() => _TenorTabViewState();
@@ -249,10 +251,10 @@ class _TenorTabViewState extends State<TenorTabView>
         fromTenor.insert(
           0,
           TenorCategory(
-            name: 'Featured',
-            searchTerm: '📈 Featured',
             image: featuredGif.media.tinyGif?.url ?? '',
+            name: widget.featuredCategory,
             path: featuredCategoryPath,
+            searchTerm: 'featured',
           ),
         );
       }