+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.0.2
 
 - chore: write tests
+- fix: category was displaying `searchTerm` and now it's displaying `name` as that localized from Tenor
+- fix: issue where featured category could not be localized
 - refactor: support `tenor_dart` changes
 
 [All Code Changes](https://github.com/Flyclops/tenor_flutter/compare/0.0.1...0.0.2)